requirements.txt
faiss_cpu==1.8.0.post1
numpy==1.23.4
pandas==2.2.2
Pillow==10.4.0
sentence_transformers==3.0.1
streamlit==1.22.0
streamlit_option_menu==0.3.13
torchvision
torch
pydub
speechrecognition
opencv-python